+++ NEWS.md
@@ -21,3 +21,5 @@
   (#122).
 
 * `%||%` is now exported (#109).
+
+* `update_list()` can now modify an element called `x` (#98).